refactor: rename folder for convention
