+++ viewer/packages/rendering/src/render-pipeline-providers/CadGeometryRenderModePipelineProvider.ts
